Fix gcc unsigned vs int warning
Date: 2025-07-26 04:31:28 UTC

+++ src/io/kdebug/kdebug.c
@@ -41,7 +41,7 @@ int
 
 	const char debug_message[] = "[DEBUG] ";
 	serial_writes(debug_message);
-	count += kstrlen(debug_message);
+	count += (int) kstrlen(debug_message);
 
 	for ( const char *p = format; *p; ++p ) {
 		if ( *p != '%' ) {
